+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.3.36 - Graph Path Problems 
+  * **Commit**: 2026-02-08 20:55
+  * problem.NewGraphPath
+  * data.GraphPath, NewGraphPath
+  * fn.AsGraphPath, StringGraphPath
+  * fn.ConstraintSimplePath
+  * fn.PathDistances
 ## v0.3.35 - K-Center Problems 
   * **Commit**: 2026-02-08 16:04
   * problem.NewKCenter